Refactor the h11 and h2 serving code
This removes the inheritance from the http module for h11 and h2 on
the grounds that it made the code more confusing and difficult to
follow. In the process it also made it clear that the code was
unnecessarily trying to control back-pressure that should rather be
left to the framework. Specifically this means that an ASGI framework
can block whilst writing, but should still receive a disconnect
message and cleanup.