Add the ability to load the configuration from a file
This allows for the command line argument ``-c`` or ``--config`` to be
used to specify a TOML or python file containing the configuration. If
a python file is used the argument should be prefixed with
``python:`` (this prefix follows Gunicorn's usage).

TOML has been chosen as a nicer choice than JSON or YAML by my
subjective view.