Bugfix ensure the event-loop is configured before the app is created
This ensures that any asyncio primitives created with the app are
aware of the correct (fully configured) event-loop. This was a problem
as the loop would be re-created after the application was loaded
leading to run time warnings. This preserves the run_single behaviour
except that providing the event loop is now mandatory in order to make
this a clear requirement for users of ``run_single``.