Add worker-class option to the config and CLI
This allows the worker class to be specified as either asyncio, or
uvloop currently but with more options planned. The naming matches
Gunicorn.

A desire to match Gunicorn's conventions also results in the -k to -w
change present in this commit, which could cause confusion initially
and is backwards incompatible.