Add initial trio worker
This only supports HTTP/1 and is not complete (no upgrade support),
however it is a good checkpoint for the introduction of a trio worker
in that it actually works as an ASGI server for HTTP/1.1.

Interestingly this allows asyncio to be benchmarked against trio, with
these local results,

███████                                       1973  Hypercorn-trio
██████████████                                3590  Hypercorn-asyncio
█████████████████████                         5388  Hypercorn-uvloop

for a simple GET request. See the quart-benchmark repository for the
methodology. Note much work must be done before these results are
trustworthy (although interesting now).

This also refactors code where possible to be shared between the
asyncio and trio workers (also allowing for better tests).

The following commits will introduce a full trio worker.