Refactor the h11 serving code
This is primarily to correct the logic regarding when to close the
connection and whether to send a disconnect message to the ASGI
app. More specifically this refactor makes the close decision the
responsibility of the worker rather than common code. In addition it
hopefully clarifies the lifecycle of the connection in terms of
requests and request-recycling. I hope as well that this makes the
code easier to follow and reason about.