Refactor the wsproto serving code
This is primarily to correct the logic regarding when to close the
connection and whether to send a disconnect message to the ASGI
app. More specifically this refactor makes the close decision the
responsibility of the worker rather than common code. I hope as well
that this makes the code easier to follow and reason about. It does
improve the autobahn test compliance, as regards to the closing tests.

This is roughly the limit without making some improvements to the
wsproto library itself, which I'll look at doing next.

The changes to the trio h11 code are too allow for the upgrade
exception to propagate without closing the connection.

There is a remaining possible issue with the close frames, which
relates to what the server should do if the client doesn't respond but
rather just holds the connection open. I will open an issue for a
websocket idle timeout.