Refactor the h2 serving code
This is primarily to correct the logic regarding when to close the
connection and whether to send a disconnect message to the ASGI
app. More specifically this refactor makes the close decision the
responsibility of the worker rather than common code. It also makes
all the IO based decisions (flow control and server push) the
responsibility of the worker code. I hope as well that this makes the
code easier to follow and reason about.

There is a remaining issue with a lack of idle timeouts in the trio
worker, but I'm unsure how to solve this at the moment.