Shutdown all tasks on closure for asyncio workers
Trio workers already have this functionality via the nursery usage.

This helps ensure the server gracefully shuts down, as it allows the
CancelledErrors to be dealt with if required. The code is based on the
asyncio.run function present in 3.7 (this itself can't be used as 3.6
is supported and as Hypercorn needs to be able to use a supplied the
event loop).

Note the lifespan worker had a CancelledError bug that this surfaced
and fixed.