Alter multi worker shutdown
This switches the shutdown prompt from a signal to based on a
multiprocessing event being set (which is polled). This should mean
that multiple workers gracefully shutdown on windows, as on windows
there is no way (I can find) to send a signal to the worker processes
and the terminate kills them outright.