Allow further h2 initial settings to be configured
This allows the max concurrent streams and max header list size
settings to be configured via the configuration object. Note that the
creation of the settings instance is required due to a h2 pecularity
with the max concurrent streams (it is hardcoded to 100).

The default values match those used in the h2 library and are
considered safe defaults.