Bugfix don't report asyncio task cancellation as ASGI framework errors
Trio seems to have made a better decision here in that the trio
CancelledError inherits from BaseException rather than Exception. In
asyncio CancelledError inherits from Exception. This means that it was
previously caught and reported as an ASGI framework error, despite
neither being an error nor the frameworks fault.

This should have no affect on the trio worker.