Bugfix ensure ASGI framework errors are logged
The lifespan part of the ASGI specification is new, and hence
Hypercorn will allow for frameworks that don't support it (the first
exception catch). However it should still show errors that occur in
the framework after the scope callable is called. Without this
debugging errors on startup (or shutdown) is very difficult.