Add an official API for using Hypercorn programmatically
It is not expected that all usage of Hypercorn will be via the command
line, in fact both Quart and Quart-Trio use Hypercorn programmatically
via unofficial functionality. This standardises the API to the following
signature,

    async def serve(app: Type[ASGIFramework], config: Config) -> None:

with the intention being that this remains unchanged as Hypercorn
evolves.

In addition the documentation has been updated to reflect this.