Ensure that h2c upgrade requests work
Previously the h2c upgrade in the asyncio worker would not be
responded to (the coroutine wasn't awaited). Whereas the trio worker
had no code at all to handle a h2c upgrade and would crash strangly.

This now works and it is possible to upgrade from HTTP/1 to HTTP/2 on
an unencrypted (clear) connection.