Support requests that assume HTTP/2
A HTTP/2 request starts with ``PRI * HTTP/2.0`` which can
intentionally be parsed by HTTP/1 parsers. When this is received
Hypercorn will now switch to the h2 mode and respond as such. This
happens when the client assumes the server supports HTTP/2 (without
consulting the ALPN or NPN flags or setting them) and sends a HTTP/2
request.