Bugfix signal handling
Signal handling with asyncio is an absolute pain. The previous version
would not gracefully shutdown. This version does by ensuring that
SIGINT or SIGTERM results in a Shutdown exception being raised (either
due to the loop handling the signal in single worker mode or the main
process doing so). The Shutdown exception can be caught and allow for
a graceful shutdown (lifespan shutdown). Note this is required due to
the changes in 1ecfcc3ac8a1da1b8b12efdd5cab51fd41ad1c2f which prevented
the try-except wrapping the loop.run_until_complete().

The run code is a basic copy of that in asyncio, and I'll switch when
Python 3.6 support is dropped.