Allow the access logger class to be customised
This also simplifies the code to access log within the various ASGI
protocol handlers. Note that the logger is created in the master
process (as with Gunicorn) and shared with the workers.

This is similar to Gunicorn in that the class of the logger can be
customised, however the mechanism requires a Python object rather than
a referenced string.