Bugfix ensure connection is not closed when reseting streams
This fixes a bug in 39041f27040edec88dd4840122812a96d443e7d9 whereby
the stream would be closed due to the client closing the
connection. In this case the stream would be unable to complete and
hence require a reset, however as the connection is closed the reset
is invalid.

It however raises the issue that in the asyncio worker these stream
tasks are destroyed without being awaited (no nursery to clean them
up as in Trio).