Support ASGI 3.0
This adds support for the latest ASGI version, whilst keeping support
for ASGI 2.0. In addition it requires the Lifespan support (and spec)
to be upgraded to 2.0.

Note this has required the clarification of the ASGIFramework types,
which results in a backwards incompatible change if you are reliant on
Hypercorn's types. On the up side this will result in mypy warnings
and hopefully introduce you to ASGI 3.0.