Support serving on insecure binds alongside secure binds
The primary usage of this is to redirect HTTP traffic to HTTPS, as
such ASGI middleware to do so has been included. Care must be taken
when using this to ensure that the HTTP traffic is not served but
redirected.

This means that there need be nothing in front of Hypercorn, HTTP to
HTTPS was the last remaining missing feature. Allowing Hypercorn to
connect directly ensures HTTP/2 is delivered to Hypercorn rather than
being terminated upstream (as Nginx will do).