Bugfix lifespan failure handling
The version implemented in d47b2033e7db0eacbd7118534a35ef7dbe3bf694
would not fail correctly in the asyncio worker, as it would crash the
task but not the server (the server would fail after the startup timed
out). To ensure both trio and asyncio work correctly on lifespan
failure the tests have been updated to actually represent how the
lifespan is called and used.

This also wraps the failure message in single quotes to help it stand
out.