Bugfix ensure early connection closure
If a keep-alive request is made but the application choses to respond
before consuming the entire body, the connection cannot be
recycled. Additionally the body may still be sent which Hypercorn
would need to receive despite the application ignoring it. Previously
Hypercorn would error trying to keep the connection alive (cannot
recycle until body has been received). The simplest solution to this
is for the connection to be closed.

Note this only works for the asyncio worker, trio requires more
investigation.