Structually refactor the codebase
This is a structural refactor aimed at reducing the amount of
duplicated logic and to separate the event loop specific code from the
protocol specific code. It will allow for another event loop (curio)
to be supported (in a following commit) and more if they exist.

The logic to the refactor is that the IO and app handling parts are
event loop specific, whereas the protocol part isn't (although it does
depend on a normalised Event). The more logic that isn't event loop
specific that can be shared the less repitition. The protocol parts
are structured as streams with a handle method (receive) and passed a
send callable.

This also revereses an earlier decision to use the asyncio callback
approach switching to a stream approach. Whilst the later is reported
slower (this may have changed in the last year) it allows the stream
structure to be used. As Uvicorn is already faster than Hypercorn, the
raw speed is likely less of a draw than the feature set.