Bugfix WebSocket shutdown
There are two parts to this, firstly the WebSocket connection won't
receive a CloseConnection event from itself, so the code handling this
can be removed. Secondly the instruction to close the connection
cannot immediately close the connection as the client needs to
respond.

There is an issue that if the client never responds with a close the
connection will remain open. This will be fixed in a subsequent
commit.