Rationalise the logging settings based on Gunicorn
This deprecates `--access-log` and `--error-log` and their associated
config names `access_logger` and `error_logger` for `--access-logfile`
and `accesslog`, and ``--error-logfile` and `errorlog`. This matches
the configuration used by Gunicorn. The originals will be removed in
0.8.

This also clarifies that the config.log can be used throughout to log,
as will be used in subsequent commits.