Allow the server header to be omitted
This adds a config option ``include_server_header`` which can be set
to False in order to prevent Hypercorn from adding the server
header. Additionally the config ``response_headers`` method can be
overidden in order to fully control headers added by Hypercorn.

This is often desired to obfuscate the server version in the hope it
improves security.