Bugfix early exit h11 event handling
If the connection state is DONE (or CLOSED or MUST_CLOSE) Hypercorn
will not expect to handle events from the client, hence it is best to
early exit.

This caused an error whereby Hypercorn responded before the request
body had been fully received. Specifically,

    File "/usr/local/lib/python3.7/site-packages/Hypercorn-0.6.0+dev-py3.7.egg/hypercorn/protocol/h11.py", line 156, in _handle_events
        await self.stream.handle(Body(stream_id=STREAM_ID, data=event.data))
    AttributeError: 'NoneType' object has no attribute 'handle'

as the stream had been set to None on completion of the response.