Bugfix correctly handle client disconnection
This ensures that when a client disconnects and the server is still
reading or writing a Closed event is sent to the protocol and the
read/write is aborted.

Note that the trio too slow handling was missleading as the client
disconnection would not trigger this, nor would the keep alive
timeout.