Bugfix change the h2 closed routine
It was possible for the streams dictionary to change size during
iteration (if a stream closed thanks to an await context switch). This
should remove this error by ensuring the stream ids list is fixed and
utilising the safer close from
f12c9c5b4d9ebf4babb1140029304c5967dbdf73.