Bugfix handle ASGI apps ending prematurely
Previously if an ASGI app errored Hypercorn would attempt to cleanup
and close the connection. However if the ASGI app completed, without
erroring, Hypercorn would not attempt to cleanup and close. This could
leave the connection in a broken state with no way for Hypercorn to
know what to do.