Support HTTP/2 prioritisation
This sends data as prioritised by the client using the Hyper priority
library to build and track the priority tree. I've chosen to have a
seperate task consume data from stream buffers and then send the data
as prioritised - this seems to be the simplest way, although it is
harder to reason about.