Add the ability to change permissions and ownerships of unix sockets
An example of how this is useful is to consider using Hypercorn with
nginx where nginx is configured to use an upstream socket. In this
case nginx is often configured to run as another user, which will not
be able to read the Hypercorn created socket. Hence telling Hypercorn
to change the ownership helps.

Note this contains a usage update for the quic binds (this was missed
previously).