Bugfix clarify lifespan error messages
If the app handling the lifespan fails before the startup the message
should make it clear there will be no lifespan support. If it fails
after startup but before shutdown it should make it clear only
shutdown failed. Otherwise it should note that the app failed after
shutdown - probably because the application failed to stop/return.