Bugfix remove run_forever usage in worker_serve
This fixes an error introduced in
836540ac4ce099ecedd7c99e5d18a92652962e55. The issue was that in the
case whereby serve was used in an event loop without a shutdown
trigger it would try to call the run_forever, which would fail as the
loop was already running. To solve this it instead awaits a Future (as
the shutdown trigger) which never completes (it is never set).

The Future 'hack' is required as Hypercorn can't await the TCP servers
as it may only be serving UDP which doesn't have a serve_forever
equivalent.