Allow the Alt-Svc headers to be configured
If specified the configured Alt-Svc headers are used in preference to
auto generated ones. This allows for port remapping whereby Hypercorn
is not aware that the service is exposed on a different port to the
one Hypercorn is serving on.

This is a provisional API, I expect for a better version to become
apparent in time.