Bugfix HTTP/1 connection recycling
If a subsequent request on a kept alive connection is received before
the http stream indicates it has closed i.e. between the stream
sending EndBody and StreamClosed the h11 protocol would recycle after
receiving the subsequent request. This meant that the request would
not be handled, the connection would stall and then timeout.

This is mostly seen in benchmarking (as the requests have to sent in
quick succession), although there are bug reports for production
issues.