Bugfix correctly cope with TCP half closes via asyncio
This introduces a simple version of the trio nursery to manage the
tasks spawned in the TCP server - this makes it much easier to reason
about and understand. It is slightly based on CoroExecutor.

The issue is that a half close (client sends `b""`) would cause the
server to close before it had finished responding i.e. the app task
should be awaited. The task group now ensures that the task finishes
before allowing the group with block to exit and the server to close.