Bugfix don't error if send to a h11 errored client
If the client is in a errored state after the request line, for
example from sending body beyond what the request headers indicated,
the asgi app will try to respond to a client in a error state. This
would previously have errored, now the error is ignore and the
connection dropped as normal.