Add support for server names
This allows the valid hosts to be configured and for Hypercorn to only
reply to these hosts (404s to others). This helps mitigate DNS
rebinding attacks.

As Hypercorn is often used without a server name validating proxy (for
HTTP/2 and HTTP/3) it makes sense for this functionality to be in
Hypercorn.

I'm unsure what the best default is, I think there is a case for the
secure default to require server names to be set. However I think this
is too restrictive to the actual use cases.