Refactor the event loop specific handling
Previously the event loop specific calls, e.g. spawn, sleep, time, all
had to be done within the xxx_server modules with the exception of the
spawn app. However this has become unweidly (e.g. the send_task or
call_at) and is making feature development, such as ping intervals
difficult. This refactor instead passes the relevant functionality to
the protocols in the context object.

With thanks to @petrikigor for the inspiration for the context object.