[h2c] add missing Connection header field
1. Without this field, nghttp2 report the following errors:

[ERROR] Failed to parse HTTP Upgrade response header: (HPE_INVALID_CONSTANT) Expected HTTP/

2. https://tools.ietf.org/html/rfc7230#section-6.7

When Upgrade is sent, the sender MUST also send a Connection header
field (Section 6.1) that contains an "upgrade" connection option.

3. https://developer.mozilla.org/en-US/docs/Web/HTTP/Protocol_upgrade_mechanism

Because Upgrade is a hop-by-hop header, it also needs to be listed in the
Connection header field.

After creating the initial HTTP/1.1 session, you need to request the upgrade
by adding to a standard request the Upgrade and Connection headers.