Handle sockets given as file descriptors properly
Currently file descriptors are converted into python sockets using
`socket.fromfd`, this method has two flaws. The first is that it
duplicates the file descriptor, this results in a dangling file
descriptor which is left behind if it is not manually closed. Secondly
it requires passing the socket type and family, of which we have no
knowledge of and are left to give some defaults which are likely to be
wrong. Instead of this, `socket.socket(fileno=fd)` can be used. This
adopts the file descriptor rather than closing it and infers the socket
parameters.