Add typing for the ASGI messages
This makes use of the TypedDict (added to Python in 3.8 and available
in the typing_extensions package) to specify the types of the key
values in the ASGI messages. This then ensures that the messages are
correctly constructued and used in the code.

Note the type ignores are due to this issue
https://github.com/python/mypy/issues/8533.