Bugfix the root_path handling
If a root_path is specified the server should return 404 to any
requests that have a path that do not start with the root_path. The
server should also adjust the path to remove the root_path, whilst
keeping the raw_path intact.

This allows the root_path to be used as a global prefix to all routes
served.