Correctly utilise SCRIPT_NAME in the wsgi middleware
If a root_path (WSGI SCRIPT_NAME) is specified the server should
return 404 to any requests that have a path that do not start with the
SCRIPT_NAME. The server should also adjust the path to remove the
SCRIPT_NAME, whilst keeping the raw_path intact.

This allows the SCRIPT_NAME to be used as a global prefix to all
routes served using the WSGI middleware.