Bugfix eof and keep alive handling
A EOF indicator, b"", sent by the client when closing (or half
closing) should be passed to the protocol for the protocol to decide
what to do. For example H11 will RemoteProtocolError if it expects
more data from the client or simply return a ConnectionClosed event if
not.

Related to this is noting that the protocol should indicate when it is
idle/or not, via the updated event, to start/stop a keep alive
timeout.

This should result in earlier disconnection events and a more robust
server.