Improve the graceful shutdown
This should ensure that Hypercorn waits for serving tasks to complete
within a graceful timeout before cancelling them. This allows for a
graceful shutdown. In order to ensure that no new requests are handled
during this graceful timeout the servers are shutdown beforehand and
the protocols reject any new requests.

For HTTP/1 the server will not recycle the connection during the
graceful timeout whereas for HTTP/2 any new requests will be reset
(rejected) and the settings changed to 0 max concurrent
connections. Note the HTTP/2 shutdown takes inspiration from
https://github.com/akka/akka-http/issues/3580

It isn't clear how to handle this gracefully for HTTP/3, so this may
not work as expected.