Bugfix only recycle a HTTP/1.1 connection if client is DONE
This will close rather than recycle a connection if the client is
still sending a body when the server is done recycling.

It is feasible for the server to wait for the client to finish sending
however this solution prevents a client slowly sending a body to the
server in an attempt to tie up server resources.