Close idle Keep-Alive connections on graceful exit
This change makes the Keep-Alive timeout routines sensitive to the
global termination switch, so that a connection in a Keep-Alive state
will immediately close when termination is requested. It achieves this
by making the termination flag on the context an event instead.

We also fix a minor bug where the Keep-Alive timeout would never be
reset after the first time it was created.

More work is needed for HTTP/3, as the semantics of a connection in a
GOAWAY state are different than what is currently exposed as "idle" from
the UDP server's perspective.