Directly support serving WSGI applications
This allows Hypercorn to serve a wsgi application via:

    hypercorn module:app

With requests being run in threads via an asyncio executor or a trio
run_sync call.

The type of app will be determined automatically, however if it isn't
an additional prefix can be used to specify wsgi or asgi as:

    hypercorn wsgi:module:app
    hypercorn asgi:module:app

Note that this drops support for ASGI2 applications as it is unlikely
this is useful anymore.