Alter the process and reloading system
Rather than using a single process with the reloader and restarting
the entire program there is now a main process that checks for changes
and restarts the process workers if so. This should resolve the
various bugs trying to restart the entire program.

A downside is that there is now always at least two processes and it
is harder to use a reloader when using the API method to serve an app.

Note the application must be loaded in the main process so that the
observer knows what files to check for changes.