Bugfix ensure that closed is sent on reading end
This ensures that the protocols receive this event event when the
socket closes unexpectedly in the asyncio case on OSX, see
https://github.com/pgjones/hypercorn/issues/50.

This requires the http_stream to allow messages to continue after
closure, but only the actual messages - not None as that indicates the
app has finished.

Thanks to @jonathanslenders, and @tohin for related solutions.