Use more modern asyncio apis
This utilises the taskgroup backport to allow usage of these apis in
pre Python 3.11 code.

This should mean the code is more reliable, and robust. It definetly
means it is clearer.

Note I've switched from WeakSet to Set and the after_done_callback as
the Python TaskGroup code uses Set. I hope this could fix and explain
a possible memory leak reported by some users.