Support restarting workers after max requests
This is useful as a "solution" to memory leaks in apps as it ensures
that after the max requests have been handled the worker will restart
hence freeing any memory leak.

The options match those used by Gunicorn.

This also ensures that the workers self-heal such that if a worker
crashes it will be restored.