Don't error on LocalProtoclErrors for ws streams
There is a race condition being hit in the autobahn compliance tests
whereby the client closes the stream and the server responds with an
acknowledgement. Whilst the server responds the app sends a message,
which now errors as the WSConnection state is closed.

As the state is managed by the WSStream, rather than the WSConnection
it makes sense to ignore these errors.