Send a 400 if data is received before the websocket is accepted
This is necessary as the data is not known to be either websocket or
http data as the server must accept or reject the connection and hence
it is a bad request.

In practice this is rare as the upgrade request is a GET request which
rarely has body data and most websocket clients wait for acceptance.